Separated Discovery into a new package.
