+++ app/src/main/java/com/speakeasy/watsonbarassistant/Discovery/OnTaskCompleted.kt
@@ -1,0 +1,8 @@
+package com.speakeasy.watsonbarassistant.Discovery
+
+import com.speakeasy.watsonbarassistant.DiscoveryRecipe
+
+
+interface OnTaskCompleted {
+    fun onTaskCompleted(recipes: MutableList<DiscoveryRecipe>)
+}
