+++ app/src/main/java/com/speakeasy/watsonbarassistant/Discovery/SearchDiscovery.kt
@@ -8,24 +8,18 @@ import com.speakeasy.watsonbarassistant.*
 import kotlinx.serialization.json.JSON
 import java.util.*
 
-class SearchDiscovery(inputListener: OnTaskCompleted) : AsyncTask<List<Ingredient>, Void, MutableList<DiscoveryRecipe>>() {
+class SearchDiscovery(private val inputListener: OnTaskCompleted):
+        AsyncTask<Array<Ingredient>, Void, MutableList<DiscoveryRecipe>>() {
 
-    private var listener: OnTaskCompleted
-
-    init {
-        this.listener = inputListener
-    }
-
-    override fun doInBackground(vararg args: List<Ingredient>): MutableList<DiscoveryRecipe> {
-        var ingredients = args[0]
-        var orderedRecipes = PriorityQueue<DiscoveryRecipe>(10, CompareLists)
-
-        Log.i("Discovery", "Start.")
-        val discovery = Discovery(
-                VERSION_DIS,
-                USERNAME_MIKE_DIS,
-                PASSWORD_MIKE_DIS
-        )
+    override fun doInBackground(vararg args: Array<Ingredient>): MutableList<DiscoveryRecipe> {
+        val ingredients = args[0]
+        val orderedRecipes = PriorityQueue<DiscoveryRecipe>(10) { a, b ->
+            when {
+                a.queueValue < (b.queueValue) -> 1
+                else -> -1
+            }
+        }
+        val discovery = Discovery(VERSION_DIS, USERNAME_MIKE_DIS, PASSWORD_MIKE_DIS)
         discovery.endPoint = URL_MIKE_DIS
 
         val queryBuilder = QueryOptions.Builder(ENV_ID_MIKE_DIS, COL_ID_MIKE_DIS)
@@ -33,46 +27,24 @@ class SearchDiscovery(inputListener: OnTaskCompleted) : AsyncTask<List<Ingredien
         queryBuilder.query(buildIngredientQuery(ingredients)).count(50)
         val queryResponse = discovery.query(queryBuilder.build()).execute()
 
-        Log.i("Discovery", "Serialization Start.")
-        //Log.i("Discovery", "Response: " + queryResponse.results[0])
-        for(response in queryResponse.results){
+        for (response in queryResponse.results) {
             val recipe = JSON.nonstrict.parse<DiscoveryRecipe>(response.toString())
-            if(recipe.imageBase64 == ""){
-                Log.i("Discovery","Failed: " + recipe.title + ": " + response.toString() + "\n")
-            }
             recipe.calculatePercentAvailable(ingredients)
             orderedRecipes.add(recipe)
         }
-
-        Log.i("Discovery", "Serialization Finished.")
-
+        orderedRecipes.filter {
+            it.title != "" && it.ingredientList.count() > 0
+        }
         return orderedRecipes.toMutableList()
-    }
-
-    protected fun onProgressUpdate(vararg progress: Int) {
     }
 
     override fun onPostExecute(result: MutableList<DiscoveryRecipe>){
         super.onPostExecute(result)
-
-        this.listener.onTaskCompleted(result, MainMenu())
+        inputListener.onTaskCompleted(result)
     }
 
-    private fun buildIngredientQuery(ingredients: List<Ingredient>): String{
-        val queryString = "ingredientList:"
-        val ingredients = ingredients
-        val sb = StringBuilder()
-        sb.append(queryString)
-
-        for (ingredient in ingredients){
-            if(ingredient.name != "") {
-                sb.append(ingredient.name)
-                sb.append("|")
-            }
-        }
-
-        sb.replace(sb.length-1,sb.length, "")
-        Log.i("Discovery", "Ingredient Query: " + sb.toString())
-        return sb.toString()
+    private fun buildIngredientQuery(ingredients: Array<Ingredient>): String{
+        return ingredients.asSequence().filter { it.name != "" }
+                .joinToString("|", "ingredientList:") { it.name }
     }
 }
